+++ catcher/cmd/commands.go
@@ -1,10 +1,12 @@
 package cmd
 
+// InitCommand - Initialize server configuration
 type InitCommand struct {
 	RetryNumber   int  `long:"retry_number" description:"Number of connection retries" required:"false" default:"5"`
 	RetryInterval uint `long:"retry_interval" description:"Period between retries in seconds" required:"false" default:"3"`
 }
 
+// RunCommand - Run server in the production mode
 type RunCommand struct {
 	ConfigurationFilename string `short:"C" long:"config" description:"Configuration filename" required:"true"`
 	Broker                string `short:"B" long:"broker" description:"Connection URL for broker" required:"false"`
