+++ catcher/cmd/init.go
@@ -1,30 +1,24 @@
 package cmd
 
 import (
-	"encoding/json"
 	"fmt"
-	"github.com/AlecAivazis/survey"
-	"io/ioutil"
 	"os"
+
+	"github.com/codex-team/hawk.catcher/catcher/configuration"
+
+	"github.com/AlecAivazis/survey"
 )
 
-type Configuration struct {
-	BrokerURL     string `json:"broker_url"`
-	Exchange      string `json:"exchange_name"`
-	Host          string `json:"server_host"`
-	Port          int    `json:"server_port"`
-	RetryNumber   int    `json:"retry_number"`
-	RetryInterval uint   `json:"retry_interval"`
-}
-
+// Default configuration parameters for the setup wizard
 var (
 	DefaultBrokerURL      = "amqp://guest:guest@localhost:5672/"
 	DefaultExchange       = "errors"
 	DefaultHost           = "localhost"
 	DefaultPort           = "3000"
-	DefaultConfigFilename = "config.json"
+	DefaultConfigFilename = "config/config.json"
 )
 
+// Configuration wizard survey
 var configurationWizard = []*survey.Question{
 	{
 		Name: "BrokerURL",
@@ -56,8 +50,10 @@ var configurationWizard = []*survey.Question{
 	},
 }
 
+// Execute initial setup
+// It creates configuration file with the help of setup wizard
 func (x *InitCommand) Execute(args []string) error {
-	var config = &Configuration{}
+	var config = &configuration.Configuration{}
 	var configFilename string
 
 	prompt := &survey.Input{
@@ -70,16 +66,19 @@ func (x *InitCommand) Execute(args []string) error {
 		return err
 	}
 
+	// if configuration file does not exist
 	if _, err := os.Stat(configFilename); !os.IsNotExist(err) {
 		overwrite := false
 		prompt := &survey.Confirm{
 			Message: fmt.Sprintf("File %s exists. Do you want to overwrite it", configFilename),
 		}
+
 		err = survey.AskOne(prompt, &overwrite, nil)
 		if err != nil {
 			return err
 		}
 
+		// exit if user does not want to overwrite file
 		if !overwrite {
 			return nil
 		}
@@ -91,36 +90,12 @@ func (x *InitCommand) Execute(args []string) error {
 		return err
 	}
 
+	// fill retry parameters from cli arguments
 	config.RetryNumber = x.RetryNumber
 	config.RetryInterval = x.RetryInterval
 
-	err = config.save(configFilename)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (config *Configuration) load(filename string) error {
-	plainText, err := ioutil.ReadFile(filename)
-	if err != nil {
-		return err
-	}
-	return json.Unmarshal(plainText, config)
-}
-
-func (config *Configuration) check() error {
-	return nil
-}
-
-func (config *Configuration) save(filename string) error {
-	configJSON, err := json.MarshalIndent(config, "", "  ")
-	if err != nil {
-		return err
-	}
-
-	err = ioutil.WriteFile(filename, configJSON, 0644)
+	// save configuration file
+	err = config.Save(configFilename)
 	if err != nil {
 		return err
 	}
