+++ catcher/main.go
@@ -1,14 +1,16 @@
 package main
 
 import (
-	"github.com/codex-team/hawk.catcher/cmd"
-	"github.com/jessevdk/go-flags"
 	"os"
+
+	"github.com/codex-team/hawk.catcher/catcher/cmd"
+
+	"github.com/jessevdk/go-flags"
 )
 
 // Command-line interface options
 var opts struct {
-	Init cmd.InitCommand `command:"init" description:"Initialize server"`
+	Init cmd.InitCommand `command:"init" description:"Initialize server configuration"`
 	Run  cmd.RunCommand  `command:"run" description:"Run server"`
 }
