+++ tools/pybomb/bomber.py
@@ -1,7 +1,7 @@
 import requests
 from datetime import datetime
 
-data = '''{"token":"eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9","sender":{"ip":"127.0.0.1"},"catcher_type":"errors/php","payload":{"title":"Field `blocks` is missing","timestamp":1545203808,"severity":16,"backtrace":[{"file":"/var/www/codex/vendor/codex-team/editor.js/EditorJS/EditorJS.php","line":77,"source code":[{"line number":76,"content":"         if (!isset($data['blocks'])) {"},{"line number":77,"content":"             throw new EditorJSException('Field `blocks` is missing');"},{"line number":78,"content":"         }"}]},{"file":"/var/www/codex/application/classes/Controller/Articles/Index.php","called line":"191","source code":[{"line number":"190","content":"     {"},{"line number":"191","content":"         $editor = new EditorJS($content, Model_Article::getEditorConfig());"},{"line number":"192","content":"         $blocks = $editor->getBlocks();"}]}],"get":{},"post":{"text":"Hello, World!","is_published":false},"headers":{},"source release":""}}'''
+data = '''{"token":"eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9","sender":{"ip":"127.0.0.1"},"catcher_type":"errors/golang","payload":{"title":"Field `blocks` is missing","timestamp":1545203808,"severity":16,"backtrace":[{"file":"/var/www/codex/vendor/codex-team/editor.js/EditorJS/EditorJS.php","line":77,"source code":[{"line number":76,"content":"         if (!isset($data['blocks'])) {"},{"line number":77,"content":"             throw new EditorJSException('Field `blocks` is missing');"},{"line number":78,"content":"         }"}]},{"file":"/var/www/codex/application/classes/Controller/Articles/Index.php","called line":"191","source code":[{"line number":"190","content":"     {"},{"line number":"191","content":"         $editor = new EditorJS($content, Model_Article::getEditorConfig());"},{"line number":"192","content":"         $blocks = $editor->getBlocks();"}]}],"get":{},"post":{"text":"Hello, World!","is_published":false},"headers":{},"source release":""}}'''
 
 
 import random
@@ -26,7 +26,7 @@ async def run(r):
     url = "http://localhost:3000/catcher"
     tasks = []
     # create instance of Semaphore
-    sem = asyncio.Semaphore(100)
+    sem = asyncio.Semaphore(50)
 
     # Create client session that will ensure we dont open new connection
     # per each request.
@@ -39,7 +39,7 @@ async def run(r):
         responses = asyncio.gather(*tasks)
         await responses
 
-number = 100
+number = 100000
 loop = asyncio.get_event_loop()
 t = datetime.now()
 future = asyncio.ensure_future(run(number))
